+++ packages/backend/src/server/api/endpoints/request-reset-password.ts
@@ -6,13 +6,15 @@
 import ms from 'ms';
 import { IsNull } from 'typeorm';
 import { Inject, Injectable } from '@nestjs/common';
-import type { PasswordResetRequestsRepository, UserProfilesRepository, UsersRepository } from '@/models/_.js';
+import type { MiMeta, PasswordResetRequestsRepository, UserProfilesRepository, UsersRepository } from '@/models/_.js';
 import { Endpoint } from '@/server/api/endpoint-base.js';
 import { IdService } from '@/core/IdService.js';
 import type { Config } from '@/config.js';
 import { DI } from '@/di-symbols.js';
 import { EmailService } from '@/core/EmailService.js';
 import { EmailI18nService } from '@/core/EmailI18nService.js';
+import { CaptchaService } from '@/core/CaptchaService.js';
+import { ApiError } from '@/server/api/error.js';
 import { L_CHARS, secureRndstr } from '@/misc/secure-rndstr.js';
 
 export const meta = {
@@ -28,7 +30,12 @@ export const meta = {
 	},
 
 	errors: {
-
+		// JUICE
+		captchaFailed: {
+			message: 'Captcha verification failed.',
+			code: 'CAPTCHA_FAILED',
+			id: 'bd16d3e5-7ca6-4f54-8e6a-cb55c187e2be',
+		},
 	},
 } as const;
 
@@ -37,6 +44,12 @@ export const paramDef = {
 	properties: {
 		username: { type: 'string' },
 		email: { type: 'string' },
+		// JUICE
+		'hcaptcha-response': { type: 'string', nullable: true },
+		'g-recaptcha-response': { type: 'string', nullable: true },
+		'm-captcha-response': { type: 'string', nullable: true },
+		'turnstile-response': { type: 'string', nullable: true },
+		'testcaptcha-response': { type: 'string', nullable: true },
 	},
 	required: ['username', 'email'],
 } as const;
@@ -46,6 +59,9 @@ export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-
 	constructor(
 		@Inject(DI.config)
 		private config: Config,
+
+		@Inject(DI.meta)
+		private serverSettings: MiMeta,
 
 		@Inject(DI.usersRepository)
 		private usersRepository: UsersRepository,
@@ -59,8 +75,20 @@ export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-
 		private idService: IdService,
 		private emailService: EmailService,
 		private emailI18nService: EmailI18nService,
+		private captchaService: CaptchaService,
 	) {
 		super(meta, paramDef, async (ps, me) => {
+			// JUICE
+			await this.captchaService.verifyRequestCaptcha(this.serverSettings, {
+				hcaptcha: ps['hcaptcha-response'],
+				mcaptcha: ps['m-captcha-response'],
+				recaptcha: ps['g-recaptcha-response'],
+				turnstile: ps['turnstile-response'],
+				testcaptcha: ps['testcaptcha-response'],
+			}).catch(err => {
+				throw new ApiError(meta.errors.captchaFailed, { message: err.message });
+			});
+
 			const user = await this.usersRepository.findOneBy({
 				usernameLower: ps.username.toLowerCase(),
 				host: IsNull(),